refactor(nerdgraph): Break Nerdgraph commands out to specific files
