+++ internal/nerdgraph/command_query_test.go
@@ -1,0 +1,16 @@
+// +build unit
+
+package nerdgraph
+
+import (
+	"testing"
+
+	"github.com/stretchr/testify/assert"
+)
+
+func TestQuery(t *testing.T) {
+	command := queryCmd
+
+	assert.Equal(t, "query", command.Name())
+	assert.True(t, command.HasFlags())
+}
